+++ packages/jest-cli/src/__tests__/cli/args.test.ts
@@ -7,6 +7,7 @@
  */
 
 import {Config} from '@jest/types';
+import {constants} from 'jest-config';
 import {check} from '../../cli/args';
 import {buildArgv} from '../../cli';
 
@@ -59,34 +60,33 @@ describe('check', () => {
     expect(() => check(argv)).not.toThrow();
   });
 
+  test.each(constants.JEST_CONFIG_EXT_ORDER.map(e => e.substring(1)))(
+    'allows using "%s" file for --config option',
+    ext => {
+      expect(() =>
+        check({config: `jest.config.${ext}`} as Config.Argv),
+      ).not.toThrow();
+      expect(() =>
+        check({config: `../test/test/my_conf.${ext}`} as Config.Argv),
+      ).not.toThrow();
+    },
+  );
+
   it('raises an exception if config is not a valid JSON string', () => {
     const argv = {config: 'x:1'} as Config.Argv;
     expect(() => check(argv)).toThrow(
-      'The --config option requires a JSON string literal, or a file path with a .js, .cjs, or .json extension',
+      'The --config option requires a JSON string literal, or a file path with one of these extensions: .js, .mjs, .cjs, .json',
     );
   });
 
-  it('raises an exception if config is not a js/cjs/json file path', () => {
-    expect(() =>
-      check({config: 'jest.config.js'} as Config.Argv),
-    ).not.toThrow();
-    expect(() =>
-      check({config: '../test/test/my_conf.js'} as Config.Argv),
-    ).not.toThrow();
-    expect(() =>
-      check({config: 'jest.config.cjs'} as Config.Argv),
-    ).not.toThrow();
-    expect(() =>
-      check({config: 'jest.config.json'} as Config.Argv),
-    ).not.toThrow();
-
+  it('raises an exception if config is not a supported file type', () => {
     const message =
-      'The --config option requires a JSON string literal, or a file path with a .js, .cjs, or .json extension';
+      'The --config option requires a JSON string literal, or a file path with one of these extensions: .js, .mjs, .cjs, .json';
 
-    expect(() => check({config: 'jest.config.mjs'} as Config.Argv)).toThrow(
+    expect(() => check({config: 'jest.configjs'} as Config.Argv)).toThrow(
       message,
     );
-    expect(() => check({config: 'jest.config.ts'} as Config.Argv)).toThrow(
+    expect(() => check({config: 'jest.config.exe'} as Config.Argv)).toThrow(
       message,
     );
   });